Cosmetic - Remove useless code.

+++ smallrye-reactive-messaging-kafka/src/test/java/io/smallrye/reactive/messaging/kafka/MapBasedConfig.java
@@ -25,7 +25,9 @@ public class MapBasedConfig implements Config {
     @SuppressWarnings("ResultOfMethodCallIgnored")
     public static void clear() {
         File out = new File("target/test-classes/META-INF/microprofile-config.properties");
-        out.delete();
+        if (out.isFile()) {
+            out.delete();
+        }
     }
 
     @Override
